backend/requirements.txt
# Django & Core (Django 4.2 LTS for compatibility)
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0

# Database & Storage
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1

# Celery (Compatible with Django 4.2)
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1

# AI & ML
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
scikit-learn==1.4.0
numpy==1.26.3
Pillow==10.2.0

# OpenAI & LLM
openai==1.10.0

# MinIO/S3
minio==7.2.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
user-agents==2.2.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0
ipython==8.20.0

# Monitoring
prometheus-client==0.19.0
